Export fp16 ONNX natively instead of post-hoc conversion
onnxconverter_common.float16 produced a graph ONNX Runtime refused to load on
DeBERTa ("Cast output (float16) does not match expected type (float)"). Trace
model.half() on CUDA instead so the fp16 graph is internally consistent.

- export_onnx.py: --fp16 now moves model + inputs to CUDA half before export
  (requires GPU); integer inputs stay integer.
- model.py: OnnxRuntimeRunner upcasts logits to float so fp16 graph output feeds
  the CPU softmax safely.
- pyproject: drop onnxconverter-common (no longer used).
- README: document the native-export rationale.

+++ sentiment-service/app/model.py
@@ -266,7 +266,8 @@ class OnnxRuntimeRunner:
             if name in self.input_names
         }
         logits = self.session.run(["logits"], ort_inputs)[0]
-        return torch.from_numpy(logits)
+        # fp16 graphs return fp16 logits; upcast so the CPU softmax downstream is safe.
+        return torch.from_numpy(logits).float()
 
 
 class TensorRTRunner:
